+++ src/middlewares/index.js
@@ -3,8 +3,8 @@ module.exports = {
   doNotWaitForEmptyEventLoop: require('./doNotWaitForEmptyEventLoop'),
   httpErrorHandler: require('./httpErrorHandler'),
   jsonBodyParser: require('./jsonBodyParser'),
-  lambdaIsWarmingUp: require('./lambdaIsWarmingUp'),
   s3KeyNormalizer: require('./s3KeyNormalizer'),
   urlEncodeBodyParser: require('./urlEncodeBodyParser'),
-  validator: require('./validator')
+  validator: require('./validator'),
+  warmup: require('./warmup')
 }
